Added support for (full + partial) word matching.
Displays the overlap and partial of two files. A set of words is generated
from each file based on delimiting whitespace. Their intersection, union,
and difference are used to compute the amount of overlap that exists.

A future implementation might support determining the amount of overlap
between an arbitrary number of files. Also, it would convenient to have a
better way to split each string, since ()'s remain attached, as do .'s
which can be kind of gross...